requirements.txt
PyMuPDF
langchain
openai
faiss-cpu
python-dotenv
streamlit
langchain
faiss-cpu 
openai
tiktoken
